SCRS-7210 Test setup fixed
