--- app/controllers/test/TestVatLodgingOfficerBuilder.scala
@@ -1,70 +1,0 @@
-package controllers.test
-
-import java.time.LocalDate
-
-import models.S4LVatLodgingOfficer
-import models.api._
-import models.view.test.TestSetup
-import models.view.vatLodgingOfficer._
-
-class TestVatLodgingOfficerBuilder {
-
-  def vatLodgingOfficerFromData(data: TestSetup): S4LVatLodgingOfficer = {
-    val homeAddress: Option[ScrsAddress] = data.officerHomeAddress.line1.map(_ =>
-      ScrsAddress(
-        line1 = data.officerHomeAddress.line1.getOrElse(""),
-        line2 = data.officerHomeAddress.line2.getOrElse(""),
-        line3 = data.officerHomeAddress.line3,
-        line4 = data.officerHomeAddress.line4,
-        postcode = data.officerHomeAddress.postcode,
-        country = data.officerHomeAddress.country))
-
-    val threeYears: Option[String] = data.officerPreviousAddress.threeYears
-    val previousAddress: Option[ScrsAddress] = data.officerPreviousAddress.line1.map(_ =>
-      ScrsAddress(
-        line1 = data.officerPreviousAddress.line1.getOrElse(""),
-        line2 = data.officerPreviousAddress.line2.getOrElse(""),
-        line3 = data.officerPreviousAddress.line3,
-        line4 = data.officerPreviousAddress.line4,
-        postcode = data.officerPreviousAddress.postcode,
-        country = data.officerPreviousAddress.country))
-
-    val dob: Option[LocalDate] = data.vatLodgingOfficer.dobDay.map(_ =>
-      LocalDate.of(
-        data.vatLodgingOfficer.dobYear.getOrElse("1900").toInt,
-        data.vatLodgingOfficer.dobMonth.getOrElse("1").toInt,
-        data.vatLodgingOfficer.dobDay.getOrElse("1").toInt))
-
-    val nino = data.vatLodgingOfficer.nino
-
-    val completionCapacity = data.vatLodgingOfficer.role.map(_ => {
-      CompletionCapacity(
-        name = Name(data.vatLodgingOfficer.firstname,
-          data.vatLodgingOfficer.othernames,
-          data.vatLodgingOfficer.surname.getOrElse("")),
-        role = data.vatLodgingOfficer.role.getOrElse(""))
-    })
-
-    val contactDetails = data.vatLodgingOfficer.email.map(_ =>
-      OfficerContactDetails(
-        email = data.vatLodgingOfficer.email,
-        mobile = data.vatLodgingOfficer.mobile,
-        tel = data.vatLodgingOfficer.phone))
-
-    val formerName = data.vatLodgingOfficer.formernameChoice.map(_ =>
-      FormerName(
-        selection = data.vatLodgingOfficer.formernameChoice.map(_.toBoolean).getOrElse(false),
-        formerName = data.vatLodgingOfficer.formername))
-
-    S4LVatLodgingOfficer(
-      previousAddress = threeYears.map(t => PreviousAddressView(t.toBoolean, previousAddress)),
-      officerHomeAddress = homeAddress.map(a => OfficerHomeAddressView(a.id, Some(a))),
-      officerDateOfBirth = dob.map(OfficerDateOfBirthView(_, completionCapacity.map(_.name))),
-      officerNino = nino.map(OfficerNinoView(_)),
-      completionCapacity = completionCapacity.map(CompletionCapacityView(_)),
-      officerContactDetails = contactDetails.map(OfficerContactDetailsView(_)),
-      formerName = formerName.map(FormerNameView(_))
-    )
-  }
-
-}
